@@ -1,58 +1,57 @@
 {
   "name": "spring-boot-seed-client",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --env=inmemory",
-    "start:open": "ng serve -o --env=inmemory",
-    "build": "ng build --env=dev",
-    "build:watch": "ng build --watch --env=dev",
-    "build:watch:inmemory": "ng build --watch --env=inmemory",
-    "build:prod": "ng build --prod --aot --env=prod",
-    "test": "ng test --sourcemaps=false",
+    "start": "ng serve --configuration=inmemory",
+    "start:open": "npm run start -- --o",
+    "build": "ng build",
+    "build:watch": "npm run build -- --watch --configuration=inmemory",
+    "build:prod": "npm run build -- --aot --configuration=production",
+    "test": "ng test",
+    "test:watch": "npm run test -- --watch",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "angular-in-memory-web-api": "0.5.4",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "angular-in-memory-web-api": "0.6.0",
+    "core-js": "^2.5.4",
+    "rxjs": "^6.0.0",
     "moment": "2.22.0",
     "jquery": "3.3.1",
     "toastr": "2.1.2",
-    "bootstrap": "3.3.7"
+    "bootstrap": "3.3.7",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2"
   }
 }
