@@ -1,4 +1,4 @@
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
 export interface CrudInterface<T> {
   getAll(): Observable<T[]>;
